feat(Link): Migrate link
